+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Changes to this project are documented in this file. More detail (including info
 
 ### Added
 - AUTOTOOL-3572: Support "use" reference for virtualServer under GSLB_Pool
+- AUTOTOOL-3655: ([GitHub Issue 688](https://github.com/F5Networks/f5-appsvcs-extension/issues/688)): Option to prefer client subnet in GSLB_Domain
 
 ### Fixed
 - AUTOTOOL-3331: Service with partial mask and custom route domain fails to POST twice